+++ yesod-core/test/YesodCoreTest.hs
@@ -32,6 +32,7 @@ import qualified YesodCoreTest.RouteAuthSubtree.Inline as RouteAuthSubtreeInline
 import qualified YesodCoreTest.RouteAuthSubtree.Nested as RouteAuthSubtreeNested
 import qualified YesodCoreTest.RouteAuthHook.Runtime as RouteAuthHook
 import qualified YesodCoreTest.RouteLeafHook.Runtime as RouteLeafHook
+import qualified YesodCoreTest.RouteLeafHook.Mounts.Runtime as RouteLeafMounts
 import qualified YesodCoreTest.RouteAuthSplit.Runtime as RouteAuthSplit
 import qualified YesodCoreTest.FallthroughMatrix.Runtime as FallthroughMatrixRuntime
 import qualified YesodCoreTest.MultiPieceNestedRuntime as MultiPieceNestedRuntime
@@ -122,6 +123,7 @@ specs = do
       RouteAuthSubtreeNested.specs
       RouteAuthHook.specs
       RouteLeafHook.specs
+      RouteLeafMounts.specs
       RouteAuthSplit.specs
       FallthroughMatrixRuntime.specs
       MultiPieceNestedRuntime.specs